Fix problem with point type compilation in dml
Previously, Point type compilation worked for ddl, but
on further testing with a live db compilation was shown
not to work in insert/select statements. Changes in this commit
include fix to Point type and elaboration on live db test.

Further tests of Point type will be needed to be sure
all behaviours are correct and to make sure all the code
that's present is in use.